Factorization of ManagerController with 4 managers : ManagerArticleController, ManagerAuthorController, ManagerCommentController, flashController
